Criar DTO-entrada

+++ src/main/java/br/com/gerenciadorDeOficina/repositories/VeiculoRepository.java
@@ -4,6 +4,6 @@ package br.com.gerenciadorDeOficina.repositories;
 import br.com.gerenciadorDeOficina.models.Veiculo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VeiculoRepository extends JpaRepository<Veiculo, Integer> {
+public interface VeiculoRepository extends JpaRepository<Veiculo, Long> {
 
 }
